LPS-113166 remove threads
